set SMIE equivalent functions in list
Rather than replacing the function itself directly, which negatively
impacts other SMIE-using modes, the function symbols are replaced in the
list that SMIE will use when indenting.

Without this change, other SMIE-using modes are affected because the
mode's setup would "setq-local" each named function-symbol but, because
these are functions' symbols, the local scope is not respected and the
global defuns are overridden.